inline TestContext into SemanticSyntaxCheckerVisitor
in preparation for actually tracking state in this visitor
